- Removed old state store implementation
